+++ QQBotService/src/main/resources/db/schema.sql
@@ -8,7 +8,7 @@ CREATE TABLE IF NOT EXISTS chat_groups (
 CREATE TABLE IF NOT EXISTS group_messages (
     id BIGINT AUTO_INCREMENT PRIMARY KEY COMMENT '自增主键',
     group_id BIGINT NOT NULL COMMENT '群号（外键关联 groups.group_id）',
-    message_id BIGINT NOT NULL COMMENT 'NapCat 消息ID',
+    message_id INT NOT NULL COMMENT 'NapCat 消息ID',
     sender_id BIGINT NOT NULL COMMENT '发送者QQ号',
     plain_text TEXT COMMENT '纯文本内容（不含CQ码）',
